// src/app/admin-view/models/marks.ts
export class Marks {
    sid: string;
    subject1: number;
    subject2: number;
    subject3: number;
    subject4: number;
    markid: number;
    tmid: string;
    total_marks: number;
    total_marks1: number;
    total_marks2: number;
    total_marks3: number;
    total_marks4: number;
    classId: number;
    subid: number;
    stud_mark: number;
}